mohamed ismail/mohamed ismail/requirements.txt
fastapi
uvicorn
PyJWT
bcrypt
pydantic
aiofiles
